sstables: Move leveled_compaction_strategy implementation to source file
It is better than putting everything in header.
